test(e2e): add test cases
- InvalidDeposit
- UnsupportedDestChain
- UnsupportedExpense
- InvalidExpense
