+++ e2e/solve/test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"math/big"
 	"net/http"
 	"time"
 
@@ -12,6 +13,7 @@ import (
 	"github.com/omni-network/omni/lib/contracts/solvernet"
 	"github.com/omni-network/omni/lib/errors"
 	"github.com/omni-network/omni/lib/ethclient/ethbackend"
+	"github.com/omni-network/omni/lib/evmchain"
 	"github.com/omni-network/omni/lib/log"
 	"github.com/omni-network/omni/lib/netconf"
 	"github.com/omni-network/omni/lib/xchain"
@@ -20,7 +22,9 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/params"
 
+	"cosmossdk.io/math"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -32,11 +36,8 @@ type TestOrder struct {
 	Expenses      solvernet.Expenses
 	Calls         solvernet.Calls
 	Deposit       solvernet.Deposit
+	ShouldReject  bool
 	RejectReason  string
-}
-
-func (o TestOrder) ShouldReject() bool {
-	return o.RejectReason != ""
 }
 
 func Test(ctx context.Context, network netconf.Network, endpoints xchain.RPCEndpoints) error {
@@ -134,6 +135,190 @@ func Test(ctx context.Context, network netconf.Network, endpoints xchain.RPCEndp
 	}
 }
 
+func makeOrders() []TestOrder {
+	users := anvil.DevAccounts()
+	var orders []TestOrder
+
+	// erc20 OMNI -> native OMNI orders
+	for i, user := range users {
+		requestAmt := math.NewInt(10).MulRaw(params.Ether).BigInt()
+
+		// make some insufficient (should reject)
+		insufficientDeposit := i%2 == 0
+		depositAmt := new(big.Int).Set(requestAmt)
+		if insufficientDeposit {
+			depositAmt = depositAmt.Div(depositAmt, big.NewInt(2))
+		}
+
+		shouldReject := insufficientDeposit
+		rejectReason := ""
+		if insufficientDeposit {
+			rejectReason = solver.RejectInsufficientDeposit.String()
+		}
+
+		orders = append(orders, TestOrder{
+			Owner:         user,
+			FillDeadline:  time.Now().Add(1 * time.Hour),
+			SourceChainID: evmchain.IDMockL1,
+			DestChainID:   evmchain.IDOmniDevnet,
+			Expenses:      nativeExpense(requestAmt),
+			Calls:         nativeTransferCall(requestAmt, user),
+			Deposit:       erc20Deposit(depositAmt, addrs.Token),
+			ShouldReject:  shouldReject,
+			RejectReason:  rejectReason,
+		})
+
+		orders = append(orders, TestOrder{
+			Owner:         user,
+			FillDeadline:  time.Now().Add(1 * time.Hour),
+			SourceChainID: evmchain.IDMockL1,
+			DestChainID:   evmchain.IDOmniDevnet,
+			Expenses:      nativeExpense(requestAmt),
+			Calls:         nativeTransferCall(requestAmt, user),
+			Deposit:       nativeDeposit(depositAmt),
+			ShouldReject:  true,
+			RejectReason:  solver.RejectInvalidDeposit.String(),
+		})
+	}
+
+	// native ETH transfers
+	for i, user := range users {
+		amt := math.NewInt(1).MulRaw(params.Ether).BigInt()
+
+		// make some under min or over max expense
+		overMax := i < 3
+		underMin := i > 6
+
+		if overMax {
+			// max is 1 ETH
+			amt = math.NewInt(2).MulRaw(params.Ether).BigInt()
+		}
+
+		if underMin {
+			// min is 0.001 ETH
+			amt = big.NewInt(1)
+		}
+
+		shouldReject := underMin || overMax
+		rejectReason := ""
+		if underMin {
+			rejectReason = solver.RejectExpenseUnderMin.String()
+		}
+		if overMax {
+			rejectReason = solver.RejectExpenseOverMax.String()
+		}
+
+		order := TestOrder{
+			Owner:         user,
+			FillDeadline:  time.Now().Add(1 * time.Hour),
+			SourceChainID: evmchain.IDMockL1,
+			DestChainID:   evmchain.IDMockL2,
+			Expenses:      nativeExpense(amt),
+			Calls:         nativeTransferCall(amt, user),
+			Deposit:       nativeDeposit(new(big.Int).Add(amt, big.NewInt(1e17))), // add enough to cover fee
+			ShouldReject:  shouldReject,
+			RejectReason:  rejectReason,
+		}
+
+		orders = append(orders, order)
+	}
+
+	// bad dest chain
+	orders = append(orders, TestOrder{
+		Owner:         users[0],
+		FillDeadline:  time.Now().Add(1 * time.Hour),
+		SourceChainID: evmchain.IDMockL1,
+		DestChainID:   1234, // invalid
+		Expenses:      nativeExpense(big.NewInt(1)),
+		Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+		Deposit:       erc20Deposit(big.NewInt(1), addrs.Token),
+		ShouldReject:  true,
+		RejectReason:  solver.RejectUnsupportedDestChain.String(),
+	})
+
+	// unsupported expense token
+	orders = append(orders, TestOrder{
+		Owner:         users[0],
+		FillDeadline:  time.Now().Add(1 * time.Hour),
+		SourceChainID: evmchain.IDMockL1,
+		DestChainID:   evmchain.IDMockL2,
+		Expenses:      unsupportedExpense(big.NewInt(1)),
+		Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+		Deposit:       erc20Deposit(big.NewInt(1), addrs.Token),
+		ShouldReject:  true,
+		RejectReason:  solver.RejectUnsupportedExpense.String(),
+	})
+
+	// invalid expense
+	orders = append(orders, TestOrder{
+		Owner:         users[0],
+		FillDeadline:  time.Now().Add(1 * time.Hour),
+		SourceChainID: evmchain.IDMockL1,
+		DestChainID:   evmchain.IDMockL2,
+		Expenses:      invalidExpense(),
+		Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+		Deposit:       erc20Deposit(big.NewInt(1), addrs.Token),
+		ShouldReject:  true,
+		RejectReason:  solver.RejectInvalidExpense.String(),
+	})
+
+	// insufficient inventory test case
+	// orders = append(orders, TestOrder{
+	//	Owner:         users[0],
+	//	FillDeadline:  time.Now().Add(1 * time.Hour),
+	//	SourceChainID: evmchain.IDMockL1,
+	//	DestChainID:   evmchain.IDMockL2,
+	//	Expenses:      nativeExpense(math.NewInt(1).MulRaw(params.GWei).BigInt()),
+	//	Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+	//	Deposit:       solvernet.Deposit{Amount: new(big.Int).Add(big.NewInt(1), big.NewInt(1e17)), Token: common.Address{}},
+	//	ShouldReject:  true,
+	//	RejectReason:  solver.RejectInsufficientInventory.String(),
+	// })
+
+	// // unsupported deposit
+	// orders = append(orders, TestOrder{
+	//	Owner:         users[0],
+	//	FillDeadline:  time.Now().Add(1 * time.Hour),
+	//	SourceChainID: evmchain.IDMockL1,
+	//	DestChainID:   evmchain.IDOmniDevnet,
+	//	Expenses:      nativeExpense(new(big.Int).Mul(big.NewInt(10), big.NewInt(1e16))),
+	//	Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+	//	Deposit:       unsupportedDeposit(big.NewInt(1)),
+	//	ShouldReject:  true,
+	//	RejectReason:  solver.RejectUnsupportedDeposit.String(),
+	// })
+
+	// // bad src chain
+	// orders = append(orders, TestOrder{
+	//	Owner:         users[0],
+	//	FillDeadline:  time.Now().Add(1 * time.Hour),
+	//	SourceChainID: 1234, // invalid
+	//	DestChainID:   evmchain.IDMockL1,
+	//	Expenses:      nativeExpense(big.NewInt(1)),
+	//	Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+	//	Deposit:       erc20Deposit(big.NewInt(1), common.Address{}),
+	//	ShouldReject:  true,
+	//	RejectReason:  solver.RejectUnsupportedSrcChain.String(),
+	// })
+
+	// same chain for src and dest
+	// orders = append(orders, TestOrder{
+	//	Owner:         users[0],
+	//	FillDeadline:  time.Now().Add(1 * time.Hour),
+	//	SourceChainID: evmchain.IDMockL1,
+	//	DestChainID:   evmchain.IDMockL1,
+	//	Expenses:      nativeExpense(big.NewInt(1)),
+	//	Calls:         nativeTransferCall(big.NewInt(1), users[0]),
+	//	Deposit:       erc20Deposit(big.NewInt(1), addrs.Token),
+	//	ShouldReject:  true,
+	//	RejectReason:  solver.RejectSameChain.String(),
+	// })
+
+	// TODO: more tests orders (different rejection cases, valid orders, etc)
+
+	return orders
+}
+
 func openAll(ctx context.Context, backends ethbackend.Backends, orders []TestOrder) (*orderTracker, error) {
 	tracker := newOrderTracker()
 
@@ -208,9 +393,9 @@ func testCheckAPI(ctx context.Context, orders []TestOrder) error {
 				return errors.Wrap(err, "decode response")
 			}
 
-			if checkResp.Rejected != order.ShouldReject() {
+			if checkResp.Rejected != order.ShouldReject {
 				return errors.New("unexpected rejection",
-					"expected", order.ShouldReject(),
+					"expected", order.ShouldReject,
 					"actual", checkResp.Rejected,
 					"reason", checkResp.RejectReason,
 					"order_idx", i,
@@ -221,7 +406,7 @@ func testCheckAPI(ctx context.Context, orders []TestOrder) error {
 				return errors.New("unexpected reject reason", "expected", order.RejectReason, "actual", checkResp.RejectReason)
 			}
 
-			if checkResp.Accepted && order.ShouldReject() {
+			if checkResp.Accepted && order.ShouldReject {
 				return errors.New("accepted but should reject")
 			}
